Move log join tests to their own file
